# frontend/openapi/return-platform.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Return Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/data-console/v1/graph/search": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Search Graph",
        "operationId": "search_graph_data_console_v1_graph_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Q"
            }
          },
          {
            "name": "expansionDepth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 0,
              "default": 1,
              "title": "Expansiondepth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/graph/nodes/{node_id}": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Get Graph Node",
        "operationId": "get_graph_node_data_console_v1_graph_nodes__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_GraphNode_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/graph/relationships/{relationship_id}": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Get Graph Relationship",
        "operationId": "get_graph_relationship_data_console_v1_graph_relationships__relationship_id__get",
        "parameters": [
          {
            "name": "relationship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Relationship Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_GraphRelationship_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/graph/nodes/{node_id}/neighborhood": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Expand Graph Neighborhood",
        "operationId": "expand_graph_neighborhood_data_console_v1_graph_nodes__node_id__neighborhood_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Node Id"
            }
          },
          {
            "name": "expansionDepth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Expansiondepth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_GraphNode_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[GraphNode]"
      },
      "APIResponse_GraphRelationship_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphRelationship"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[GraphRelationship]"
      },
...
      "FreshnessStatus": {
        "type": "string",
        "enum": [
          "LIVE",
          "CACHED",
          "STALE"
        ],
        "title": "FreshnessStatus"
      },
      "GraphExpansionLimit": {
        "properties": {
          "maxNodes": {
            "type": "integer",
            "title": "Maxnodes"
          },
          "maxRelationships": {
            "type": "integer",
            "title": "Maxrelationships"
          },
          "maxDepth": {
            "type": "integer",
            "title": "Maxdepth"
          },
          "expansionLimit": {
            "type": "integer",
            "title": "Expansionlimit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "maxNodes",
          "maxRelationships",
          "maxDepth",
          "expansionLimit"
        ],
        "title": "GraphExpansionLimit",
        "description": "Metadata about graph expansion limits."
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "provenance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "ownership": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownership"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "labels",
          "properties"
        ],
        "title": "GraphNode",
        "description": "Canonical representation of a Graph Node."
      },
      "GraphRelationship": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "startNodeId": {
            "type": "string",
            "title": "Startnodeid"
          },
          "endNodeId": {
            "type": "string",
            "title": "Endnodeid"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "startNodeId",
          "endNodeId",
          "properties"
        ],
        "title": "GraphRelationship",
        "description": "Canonical representation of a Graph Relationship."
      },
      "GraphSearchMeta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "pattern": "^\\d+\\.\\d+$",
            "title": "Schema Version",
            "default": "1.0"
          },
          "request_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus",
            "default": "LIVE"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial",
            "default": false
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningMeta"
            },
            "type": "array",
            "title": "Warnings"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphExpansionLimit"
              },
              {
                "type": "null"
              }
            ]
          },
          "isTruncated": {
            "type": "boolean",
            "title": "Istruncated",
            "default": false
          },
          "isPartial": {
            "type": "boolean",
            "title": "Ispartial",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "GraphSearchMeta",
        "description": "Metadata for Graph Search Results."
      },
      "GraphSearchResult": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GraphSearchResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/GraphSearchMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "GraphSearchResult",
        "description": "Top-level Graph Search Result."
      },
      "GraphSearchResultData": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/GraphRelationship"
            },
            "type": "array",
            "title": "Relationships"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "relationships"
        ],
        "title": "GraphSearchResultData",
        "description": "Payload for Graph Search Results."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PageMeta"
      },
      "ResponseMeta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "pattern": "^\\d+\\.\\d+$",
            "title": "Schema Version",
            "default": "1.0"
          },
          "request_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus",
            "default": "LIVE"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial",
            "default": false
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningMeta"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WarningMeta": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Source"
          },
          "code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "code",
          "message"
        ],
        "title": "WarningMeta"
      },
...
    }
  }
}